+++ apps/native-component-list/src/screens/UI/BottomSheetScreen.android.tsx
@@ -1,16 +1,27 @@
+import { Button, BottomSheet } from '@expo/ui/jetpack-compose';
 import * as React from 'react';
-import { ScrollView, Text, View } from 'react-native';
+import { ScrollView, Text } from 'react-native';
 
 export default function BottomSheetScreen() {
-  return (
-    <ScrollView style={{ flex: 1 }}>
-      <View style={{ flex: 1, justifyContent: 'center', alignItems: 'center' }}>
-        <Text style={{ textAlign: 'center' }}>Not implemented yet on Android</Text>
-      </View>
-    </ScrollView>
-  );
+    const [isOpened, setIsOpened] = React.useState<boolean>(false);
+
+    return (
+        <ScrollView
+            contentContainerStyle={{
+                flex: 1,
+                justifyContent: 'flex-start',
+                alignItems: 'flex-start',
+                padding: 8,
+            }}>
+            <Button onPress={() => setIsOpened((h) => !h)}>Toggle</Button>
+            <Text>isOpened: {isOpened ? 'yes' : 'no'}</Text>
+            <BottomSheet isOpened={isOpened} onIsOpenedChange={(e) => setIsOpened(e)}>
+                <Text>Hello world</Text>
+            </BottomSheet>
+        </ScrollView>
+    );
 }
 
 BottomSheetScreen.navigationOptions = {
-  title: 'BottomSheet',
+    title: 'BottomSheet',
 };